requirements.txt
alembic==1.16.5
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
certifi==2025.10.5
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.120.4
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.3
mdurl==0.1.2
orjson==3.11.4
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.3
sentry-sdk==2.43.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.49.3
tomli==2.4.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
